refactor: remove dead/duplicate code, extract shared utilities (v0.14.4)
Dead code removed:
- Delete HelloWorldView.tsx, Sidebar.tsx, NameScreen.tsx (orphaned components)
- Delete scroll-area.tsx (unused UI primitive), dashboard/ (empty dir)
- Remove dead handleVoice from GenesisFlow.tsx
- Remove dead formatTitle from LensTimeline.tsx
- Remove unused BrowserWindow imports from ViewDiscovery.ts and lens.ts
- Un-export loadSdk from SdkLoader.ts (internal only)

Duplicate code consolidated:
- Extract formatTitle() and formatDisplayValue() to utils.ts (was 6 identical copies)
- Replace local formatTitle/formatCell/formatValue in all Lens views
- Extract findSystemNode to shared nodeResolver.ts (was 2 copies in SdkLoader + idea adapter)

Unused deps:
- Remove autoprefixer (not needed with Tailwind v4 + @tailwindcss/postcss)
