+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "chamber",
   "productName": "chamber",
-  "version": "0.14.3",
+  "version": "0.14.4",
   "description": "Genesis Mind Interface — desktop chat UI for Genesis agents",
   "main": ".vite/build/main.js",
   "private": true,
@@ -42,7 +42,6 @@
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.62.0",
     "@vitest/coverage-v8": "^4.1.4",
-    "autoprefixer": "^10.4.27",
     "electron": "41.2.0",
     "eslint": "^8.57.1",
     "eslint-plugin-import": "^2.32.0",
